implementation of modular serializer/deserializer
