+++ src/main/java/it/auties/whatsapp/serialization/ControllerSerializerProvider.java
@@ -7,7 +7,7 @@ import it.auties.whatsapp.controller.Store;
  * This interface provides a standardized way to serialize a session. Implement this interface and
  * <a href="https://www.baeldung.com/java-spi#3-service-provider">register it in your manifest</a>
  */
-public non-sealed interface ControllerSerializerProvider extends ControllerProvider {
+public interface ControllerSerializerProvider {
     /**
      * Serializes the keys
      *
